// src/app/api/model/statistiquesNiveauResponse.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StatistiquesNiveauResponse { 
    niveauId?: number;
    nomNiveau?: string;
    nombreClasses?: number;
    nombreEleves?: number;
    nombreLivres?: number;
    pointsMoyens?: number;
}