// src/app/api/model/statistiquesClasseResponse.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StatistiquesClasseResponse { 
    classeId?: number;
    nomClasse?: string;
    niveau?: string;
    nombreEleves?: number;
    pointsMoyens?: number;
}